[package]: react, nextjs version up (#112)
