+++ app/(landing)/layout.module.scss
@@ -1,4 +1,4 @@
-@use '/styles/color.scss' as color;
+@use '@/styles/color.scss' as color;
 
 .container {
   display: flex;
