+++ app/(main)/graduation-progress/[studentCode]/[department]/layout.tsx
@@ -1,16 +1,22 @@
-'use client';
+'use client';;
+import { use } from "react";
 
 import { TopNavigation } from '@/components/ui/TopNavigation';
 import { useInternalRouter } from '@/hooks/useInternalRouter';
 import styles from './layout.module.scss';
 
-export default function GraduationProgressLayout({
-  children,
-  params,
-}: {
-  children: React.ReactNode;
-  params: { studentCode: string; department: string };
-}) {
+export default function GraduationProgressLayout(
+  props: {
+    children: React.ReactNode;
+    params: Promise<{ studentCode: string; department: string }>;
+  }
+) {
+  const params = use(props.params);
+
+  const {
+    children
+  } = props;
+
   const router = useInternalRouter();
   const { studentCode, department } = params;
   const decodedMajorName = decodeURIComponent(department);
