+++ components/ui/TopNavigation/TopNavigation.module.scss
@@ -1,5 +1,5 @@
-@use '/styles/typography.scss' as typography;
-@use '/styles/color.scss' as color;
+@use '@/styles/typography.scss' as typography;
+@use '@/styles/color.scss' as color;
 
 .container {
   display: flex;
