+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "eslint .",
@@ -30,24 +30,24 @@
     "date-fns": "^4.1.0",
     "iron-session": "^8.0.4",
     "lottie-react": "^2.4.0",
-    "next": "14.2.10",
+    "next": "15.2.4",
     "playwright": "^1.49.1",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "uuid": "^11.0.3"
   },
   "devDependencies": {
     "@svgr/webpack": "^8.1.0",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "19.1.0",
+    "@types/react-dom": "19.1.1",
     "@typescript-eslint/eslint-plugin": "^8.14.0",
     "@typescript-eslint/parser": "^8.14.0",
     "autoprefixer": "^10.4.20",
     "dotenv-cli": "^8.0.0",
     "eslint": "^8",
-    "eslint-config-next": "14.2.10",
+    "eslint-config-next": "15.2.4",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.31.0",
     "jscodeshift": "^17.3.0",
@@ -58,5 +58,9 @@
     "typescript": "5.6.3",
     "vitest": "^3.0.5"
   },
-  "packageManager": "yarn@1.22.22"
+  "packageManager": "yarn@1.22.22",
+  "resolutions": {
+    "@types/react": "19.1.0",
+    "@types/react-dom": "19.1.1"
+  }
 }
