+++ styles/global.scss
@@ -1,3 +1,3 @@
-@use 'styles/global/reset';
-@use 'styles/global/typography';
-@use 'styles/global/spacing';
+@use '@/styles/global/reset';
+@use '@/styles/global/typography';
+@use '@/styles/global/spacing';
